(gtk 1.1x) #833: "Last scrape/announce at" shows wrong time

+++ gtk/details.c
@@ -1139,7 +1139,7 @@ refresh_time_lb( GtkWidget * l, time_t t )
     if( !t )
         gtk_label_set_text( GTK_LABEL( l ), never );
     else {
-        char * str = rfc822date( t * 1000 );
+        char * str = rfc822date( (guint64)t * 1000 );
         gtk_label_set_text( GTK_LABEL( l ), str );
         g_free( str );
     }
